Rename: 폴더명및 컴포넌트 이름수정
